feat(cards): add authentication and authorization to cards module

+++ src/modules/cards/cards-controller.ts
@@ -1,5 +1,6 @@
 import { Request, Response } from "express";
 import { cards } from "../../db/schema";
+import { AuthorizationError } from "../../errors/authorization-error";
 import { CreateCardService } from "./services/create-card-service";
 import { DeleteCardService } from "./services/delete-card-service";
 import { GetCardService } from "./services/get-card-service";
@@ -10,6 +11,9 @@ type Card = typeof cards.$inferSelect;
 
 export class CardsController {
   create = async (req: Request, res: Response) => {
+    const { userPermissions } = req;
+    if (!userPermissions || !userPermissions.includes("cards:create"))
+      throw new AuthorizationError();
     const id = await new CreateCardService().execute(req.body);
     res.json({ success: true, message: "Card created.", data: id });
   };
@@ -26,11 +30,17 @@ export class CardsController {
   };
 
   update = async (req: Request, res: Response) => {
+    const { userPermissions } = req;
+    if (!userPermissions || !userPermissions.includes("cards:update"))
+      throw new AuthorizationError();
     await new UpdateCardService().execute(req.body);
     res.json({ success: true, message: "Card updated." });
   };
 
   delete = async (req: Request, res: Response) => {
+    const { userPermissions } = req;
+    if (!userPermissions || !userPermissions.includes("cards:delete"))
+      throw new AuthorizationError();
     await new DeleteCardService().execute(req.query.id);
     res.json({ success: true, message: "Card deleted." });
   };
